refactored the technical debt from yesterday, further refactoring necessary
