+++ src/bin/src/main.rs
@@ -74,7 +74,7 @@ async fn main() -> Result<()> {
     let mut cmd = Cli::command();
     if let Some(generator) = cli.generator {
         let bin_name = cmd.get_name().to_string();
-        log::info!("Generating completion file for {:?}...", generator);
+        eprintln!("Generating completion file for {:?}...", generator);
         generate(generator, &mut cmd, bin_name, &mut std::io::stdout());
         Ok(())
     } else if let Some(command) = cli.command {
