--- src/lib/src/cli/address/mod.rs
@@ -1,6 +1,0 @@
-pub mod install;
-pub use install::*;
-pub mod uninstall;
-pub use uninstall::*;
-pub mod setup;
-pub use setup::*;
