+++ src/lib/src/cli/mod.rs
@@ -1,10 +1,4 @@
 pub mod node;
 pub use node::*;
-pub mod wallet;
-pub use wallet::*;
 pub mod update;
 pub use update::*;
-pub mod bech32;
-pub use bech32::*;
-pub mod address;
-pub use address::*;
