+++ src/lib/src/cli/node/mod.rs
@@ -1,10 +1,4 @@
-pub mod install;
-pub use install::*;
-pub mod uninstall;
-pub use uninstall::*;
 pub mod run;
 pub use run::*;
-pub mod setup;
-pub use setup::*;
 pub mod sync;
 pub use sync::*;
