+++ src/lib/src/configs/globals.rs
@@ -12,20 +12,6 @@ pub const CARDANO_NODE_VERSION: &str = "1.35.0";
 pub const LD_LIBRARY_PATH: &str = "/usr/local/lib:$LD_LIBRARY_PATH";
 pub const PKG_CONFIG_PATH: &str = "/usr/local/lib/pkgconfig:$PKG_CONFIG_PATH";
 pub const CARDANO_CONFIG_FILE_NAME: &str = "cardano.toml";
-pub const CARDANO_WALLET_URL: &str =
-    "https://github.com/input-output-hk/cardano-wallet.git";
-pub const CARDANO_ADDRESS_URL: &str =
-    "https://github.com/input-output-hk/cardano-addresses.git";
-pub const CARDANO_BECH32_URL: &str =
-    "https://github.com/input-output-hk/bech32.git";
-pub const CARDANO_NODE_URL: &str =
-    "https://github.com/input-output-hk/cardano-node.git";
-pub const CARDANO_NODE_RELEASE_URL: &str =
-    "https://api.github.com/repos/input-output-hk/cardano-node/releases/latest";
-pub const CARDANO_WALLET_RELEASE_URL: &str = "https://api.github.com/repos/input-output-hk/cardano-wallet/releases/latest";
-pub const CARDANO_ADDRESS_RELEASE_URL: &str = "https://api.github.com/repos/input-output-hk/cardano-addresses/releases/latest";
-pub const CARDANO_BECH32_RELEASE_URL: &str =
-    "https://api.github.com/repos/input-output-hk/bech32/releases/latest";
 pub const LIBSODIUM_URL: &str =
     "https://github.com/input-output-hk/libsodium.git";
 pub const SECP256K1_URL: &str = "https://github.com/bitcoin-core/secp256k1.git";
@@ -47,8 +33,6 @@ pub const CONFIG_FILES: [&str; 5] = [
     "alonzo-genesis",
     "topology",
 ];
-pub const SPINNERS: [&str; 6] =
-    ["▹▹▹▹▹", "▸▹▹▹▹", "▹▸▹▹▹", "▹▹▸▹▹", "▹▹▹▸▹", "▹▹▹▹▸"];
 pub const DIRECTORIES: [&str; 11] = [
     "work",
     "ipc",
